# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TaskEarn Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [auth]
      summary: Verify email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailInput"
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── QUIZ ────────────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get available tasks
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/start:
    post:
      operationId: startTask
      tags: [tasks]
      summary: Start a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTask"
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task and earn reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompletionInput"
      responses:
        "200":
          description: Task completed, reward credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletionResult"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        referralCode:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
    VerifyEmailInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, level, balance, pendingEarnings, totalWithdrawn, referralCode, referralCount, tasksCompleted, isEmailVerified, isAdmin, quizCompleted, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        avatar:
          type: ["string", "null"]
        level:
          type: integer
        levelName:
          type: string
        balance:
          type: number
        pendingEarnings:
          type: number
        totalWithdrawn:
          type: number
        referralCode:
          type: string
        referralCount:
          type: integer
        tasksCompleted:
          type: integer
        isEmailVerified:
          type: boolean
        isAdmin:
          type: boolean
        quizCompleted:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string

    # ── Quiz ──────────────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, title, description, category, reward, estimatedMinutes, difficulty, minLevel, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        reward:
          type: number
        estimatedMinutes:
          type: integer
        difficulty:
          type: string
        minLevel:
          type: integer
        isActive:
          type: boolean
        completedByUser:
          type: boolean
        instructions:
          type: ["string", "null"]
...
    UserTask:
      type: object
      required: [id, taskId, userId, status, startedAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        status:
          type: string
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
    TaskCompletionInput:
      type: object
      properties:
        proof:
          type: ["string", "null"]
    TaskCompletionResult:
      type: object
      required: [success, rewardEarned, message]
      properties:
        success:
          type: boolean
        rewardEarned:
          type: number
        message:
          type: string
        newBalance:
          type: number
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string

    # ── Leaderboard ───────────────────────────────────────────────────────────
...