# lib/api-zod/src/generated/types/taskCompletionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TaskEarn Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskCompletionInput {
  /** @nullable */
  proof?: string | null;
}