# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TaskEarn Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  level: number;
  levelName?: string;
  balance: number;
  pendingEarnings: number;
  totalWithdrawn: number;
  referralCode: string;
  referralCount: number;
  tasksCompleted: number;
  isEmailVerified: boolean;
  isAdmin: boolean;
  quizCompleted: boolean;
  isBanned?: boolean;
  createdAt: string;
}